refactor: move sanitize-html-content from domain to application/helpers
Moved sanitize-html-content.ts and its spec from artifacts/domain/
to artifacts/application/helpers/ since it's a utility helper used
by application use-cases rather than a core domain concept.

Updated all imports in:
- create-artifact use-case
- update-artifact use-case
- revert-artifact use-case
- html-document-export service

Updated SUMMARY.md to reflect new file location.
