+++ ayunis-core-backend/src/domain/artifacts/application/use-cases/update-artifact/update-artifact.use-case.ts
@@ -10,7 +10,7 @@ import {
 import { ArtifactVersion } from '../../../domain/artifact-version.entity';
 import { AuthorType } from '../../../domain/value-objects/author-type.enum';
 import { ContextService } from 'src/common/context/services/context.service';
-import { sanitizeHtmlContent } from '../../../domain/sanitize-html-content';
+import { sanitizeHtmlContent } from '../../helpers/sanitize-html-content';
 
 const MAX_VERSION_RETRIES = 3;
